Revert "Tasks global state management"
